Add host name checks
