+++ driver/src/test/resources/ip-san-only.crt
@@ -1,0 +1,31 @@
+-----BEGIN CERTIFICATE-----
+MIIFWjCCA0KgAwIBAgIJAIZTeihiALKcMA0GCSqGSIb3DQEBCwUAMGIxCzAJBgNV
+BAYTAlVTMQswCQYDVQQIDAJDQTEQMA4GA1UEBwwHUGhvZW5peDEVMBMGA1UECgwM
+V2lkZ2V0IFdvcmtzMR0wGwYDVQQDDBR3d3cud2lkZ2V0LXdvcmtzLmNvbTAeFw0y
+MDEwMjYxNjE0MjBaFw0zMDEwMjQxNjE0MjBaMGIxCzAJBgNVBAYTAlVTMQswCQYD
+VQQIDAJDQTEQMA4GA1UEBwwHUGhvZW5peDEVMBMGA1UECgwMV2lkZ2V0IFdvcmtz
+MR0wGwYDVQQDDBR3d3cud2lkZ2V0LXdvcmtzLmNvbTCCAiIwDQYJKoZIhvcNAQEB
+BQADggIPADCCAgoCggIBANAAmdgOZLzU5XkXcDlEsgOJCHPOOngrr68SBo3h1+bz
+UXNObEUsk8dcsBtjpAdGMZQtYU8hohPIRb6VeRWuWZhF6P0XePK3m92ehwwra6rf
+IRtG+drxMxA0SXMu3O0xxMhbp7gxlgOm9rz568hsssxJnX3vGiCnwhHRuVvLVK87
+YsMe9qAHv6akq8jQOy9KDrAUQ96fdA4PkQMob/8exo1fgXpO+8Laq+7+D+sPV/n+
+oOqgeibHgjcye6gli7xIvHESyRhDSfVsS8XUUJenr6BWrv3Nqw76YadP+sXpRK0X
+JgD2kMkMX+9NipslDnTHV7yJPNvm8CpWECzJam9tbygBlGJwnXWtxOdaXFvw2qOn
+fVpnqUays403Sd9a2iXtSAx5HkaSuK1OpRJAILItbP6VcBk11174aiXGaDCd0qh/
+F0qARstiw0pNkJWqsX00WV14hsCWJffwArewn5H6lExJMe3Ea8NztnNGPltuTv3u
+UHSw7DZmVh9n55dpkEyxM/JgADB8c/ZfCrH+BqStdXjzzqGr1sJfLT1tIfV7QzqD
+x62ya/mL49AYuKbi+XT1NbvHlUQ8GvFZqOegMPz+RvkOxUdCOmQDpXf/5Vg8GZaK
+sB76nqFjLSMWZAFxaaHe45hiUJuB5kV0poNlNGC+XpXPJWik89NsXi1HBFaBksFN
+AgMBAAGjEzARMA8GA1UdEQQIMAaHBH8AAAEwDQYJKoZIhvcNAQELBQADggIBAFyQ
+mAUDljQZ6CW6WgKMkYrBQkIlCSToN5NGHh/FqkXN8bnkkTPvddoyPlqnK6RqcuAC
+uOhKTdSmIsdwSCzKwZ8mEB7yb66qF5ABjjVEQEJUAGUOlYbqqurxg1T7XXsV7rTj
+m/POa4mD8j0Qc1av3q63vv6XkVbA46u9FK6DFEmFQ+jZxIg4/iWitMP9WQicn/2I
+WExzUxjT2s3ELRL6PhFH1UZboSt3fvN9J53E3u85/yOSSuQeEiQnOH7AcQQ60tmE
+I20bx55P49nCpYj8ZxxdVKxK5D2qbE4yttpUaU2g50o9oTgV82RjMtSA3NSHzltE
+3KKoYcZRriwkhY9dFbxXy/ovfS8PNUFwLrvf3Xb6j5QFN7I3mQNAypaH6843IQYN
+HuQEoSImR/xDz/TNuIxIT1egTgS3/NoFeEJ6+wGQ6pxKY1SrmHAtIppfY92n/f7U
+P2Uwuv1gTYiDegP3yaxJ3GAddE+a61N0545Kt2A5xzBlDXtFeF5Jp3eY0HTv5qHk
+AP6wvTbroY17ADBsPfKuB55fLmWnZgt8bnp/FZct42hxfSrZN0Q2idI4eSW7FGP8
+B3kXUnPj/Mhcr35JEU73lZq6XjguSlydQN/cJuzHNm3ohniwg77uBM54+4GesSed
+AFU3AxB7rpOycEAHjDE9eJr0tczSPX3p2v66AJnF
+-----END CERTIFICATE-----
