+++ driver/src/test/resources/multiple-cns-rev.crt
@@ -1,0 +1,32 @@
+-----BEGIN CERTIFICATE-----
+MIIFezCCA2OgAwIBAgIJAJALXiBZHCVeMA0GCSqGSIb3DQEBCwUAMH0xGTAXBgNV
+BAMMEHdpZGdldC13b3Jrcy5jb20xCzAJBgNVBAYTAlVTMQswCQYDVQQIDAJDQTEQ
+MA4GA1UEBwwHUGhvZW5peDEVMBMGA1UECgwMV2lkZ2V0IFdvcmtzMR0wGwYDVQQD
+DBR3d3cud2lkZ2V0LXdvcmtzLmNvbTAeFw0yMDEwMjQwMzE3NDZaFw0zMDEwMjIw
+MzE3NDZaMH0xGTAXBgNVBAMMEHdpZGdldC13b3Jrcy5jb20xCzAJBgNVBAYTAlVT
+MQswCQYDVQQIDAJDQTEQMA4GA1UEBwwHUGhvZW5peDEVMBMGA1UECgwMV2lkZ2V0
+IFdvcmtzMR0wGwYDVQQDDBR3d3cud2lkZ2V0LXdvcmtzLmNvbTCCAiIwDQYJKoZI
+hvcNAQEBBQADggIPADCCAgoCggIBAOzQYCtUrXmdMsRMBwZkj8xBkOSB8GowYKe4
+98IM4xj1JOpKLGEZOxkEYGhOuo9fMLWmk8QNHUb5/0txQTeVLjs6O9a3aOuTJC1U
+WgYWbpP61tmftIi6453k0G3rgJgWwdQ05H816CqgBToKCizN9YwS1gFFWYkJeVcT
+WFwjNxODcOcQGPvnLu7/o9NRabJPeQwtpdzW1qmUx4ei6Q5eiMDhBo5tlwWPtSse
+B0M7Y86+j1pvamsAGFtETiC/fyGUiW8La8wrJ8Ze2CAlq5ADKghavJBDsQBy0qU7
+qxZ7OF/zl5l3cdWTQNp8BXFZ4taLPkZ8Nwpw9N5ifQF9S8hi6YGvWqYjcyk2CR2a
+BaT7ixSDWq97ZUVaxddFEy6m2TVyCpJKX8kZfkQ0cZy6kFJKvBMbI5N893jY5vw/
+1kouDi2fE2slgM/y+5gcdQCXXvrSo+DML+kWYj8h6edDl6vEXtkrccGWtPbgkIzj
+ARja3v0mTv5kmB8Si3wzXINcYMimJEvXZAx7VhbBXAtFgxyuAb39OV9fg2zwPzf0
+KLw4VVdKMVvY4BK5RQuMs7fvY1DyJaZGMz/f7y20dO9qaKq6COQ9FiXX9Bqr916S
+6+5I5v1afsmW7GVqy+aWOod0mo0MKxlXdAMeye+q4MljsxRnM28sB27hyViMt7do
+VjJeX2KNAgMBAAEwDQYJKoZIhvcNAQELBQADggIBAF7/S4DWksCjKxYBJju2qyt2
+bbkI6g130WqEfQV2s4IRfPlkbU93QN8MFoEYjuispvTWm5nW32bR4yxW2LqmVOaX
+S2GXNp8yX4FMsLgM3Y5H21ETjTTt1ZjqvZEhNvB16TpgfUrCurNxMaL8tXVDVEPv
++f2SEpeTUPKxxihHaXX0rYvvXGg5LY8TdIRca7yWoTXP3A1g6//V9oKcEvPt1U1H
+V0oisO75BNpyoVo1U4kMCjGS4Pev7R9OJuAKy+kNXm/TJce+YG6jDYoESpExTbKW
+7seV62ixfOGEqRqwcw7xeC7iKRdupCTAmXSHQrKIV5ejSvNxkwiXq5MiMZuquttb
+hVt2YKSUPDnksQyoCa7ZpoTcsy0zogVsv7icXPhpwyB8zPsDr7k0kYIfcp3gUJ/Y
+a3fA0e0SVk0Wcl/hIUZKrYU/BIhL8Y753eBDOFOfzraqDzrveiDitYo+TMvaCj6Y
+wF+T5Uj6AV4JB/KRrXwJxbBNCXhxjohCFLMlMXlCri9bvvMzuguouUJbLAqBQmF4
+1KKs5qo7YRYWIpNyMuXVp+m9ihoJvNx/pzxd055sC9gnsin1z9qZQYvE59StBagV
+oU9dHZ1qozIiJWDQYw5SiAmJmwMsfp/INGOYqbV6MD8JCpSiiTiFWCXTBV0KeUUF
+8ooGp6V2rWxNQ1GDEPbR
+-----END CERTIFICATE-----
