+++ driver/src/test/resources/multiple-cns.crt
@@ -1,0 +1,32 @@
+-----BEGIN CERTIFICATE-----
+MIIFezCCA2OgAwIBAgIJAJv7lv9v+ZtNMA0GCSqGSIb3DQEBCwUAMH0xHTAbBgNV
+BAMMFHd3dy53aWRnZXQtd29ya3MuY29tMQswCQYDVQQGEwJVUzELMAkGA1UECAwC
+Q0ExEDAOBgNVBAcMB1Bob2VuaXgxFTATBgNVBAoMDFdpZGdldCBXb3JrczEZMBcG
+A1UEAwwQd2lkZ2V0LXdvcmtzLmNvbTAeFw0yMDEwMjQwMjQ4MjVaFw0zMDEwMjIw
+MjQ4MjVaMH0xHTAbBgNVBAMMFHd3dy53aWRnZXQtd29ya3MuY29tMQswCQYDVQQG
+EwJVUzELMAkGA1UECAwCQ0ExEDAOBgNVBAcMB1Bob2VuaXgxFTATBgNVBAoMDFdp
+ZGdldCBXb3JrczEZMBcGA1UEAwwQd2lkZ2V0LXdvcmtzLmNvbTCCAiIwDQYJKoZI
+hvcNAQEBBQADggIPADCCAgoCggIBANoWZyB1LG80qwM3bF7AAsNh6tz8bAZjkoGw
+UiIcG41IfGT1rIOhr9cylnFL8ahi+rtiMOctxdnYZykp7dQ6jFgt+f3k7mxurbv6
+gBWzXSVxcn4nrwlUdCY/sonk3NcAEW1EmkgqTUDwvNS6tXXeVpPK3+NKYfIvQUyP
+hD5kbt1GwwhBZLbVqgFyeYrogLkNrH1wof7gCbhcvcrNz//nX50vEqn0B2x4XYoE
+JXzjt5+fa/JGHA6+2PsWkuhignd+BMgKZBzclhrmUYQ8CP0feguCi9pHe0xMjt1Z
+2shBKV3UMBXVnkBAR01zVZbbuIfCDWTt5POvY9T7MrLjBkvwGhKBDf3fhxTMLqpY
+C0z4w4YDk12Loy/+GFIjDZptS1TIhGkYaNyd8GuHCOFeFtnFaWnWfM6rE3y7/qaj
+XxtVb5awsopu3OeB2tlXvLRGW+YpvHlBTF0q0iqa+YnoUcJYB3FOkiKmwzt9+23/
+kJX9c//w8xM5vRV/AIsox7jSMeoeA3pnKSjHROWluM2l2G01f5rM4OUxFyrYAkys
+G83Avmg8nDVfQmxmZF0IX5O3hPsbvQyWOVUxI0e6E81MEfEyYn1Iami8BcTxYan8
+d2SloWltZcS3+tPpCQeeXUasiuVIuigo8fQ8gf4KX5/OLNHgD3DrsqRR7fjrBjRN
+dBoW8F5/AgMBAAEwDQYJKoZIhvcNAQELBQADggIBADMPzwooMW48grkAKSESTn0f
+FURFa8CULK1hThU9VT52aGpjZSD05l5I/60ixqvOD6Gi7S5sU63XPMyfHSUSCcQ+
+MQTSnflIYcV5CTZKyFq1hfnaz15KIXxryHs6472ttkoVv2uiWGcaVyrVEf18e6bB
+IAbGSNzeH39XdLGOtdKMx81wZXdAbvDVjY3IBQVp/VlUvjNBgxmHxLUmu9jr6n66
+HNR4bn7TUlZAmkkxhlC6WcSm1XPRxdP1eLDg/r7K2a6G4qCJ5lJbDl6gNQnYO5ko
+hTDHIR2vp/nJJiuLEUF76l6dFtsiRBTraLZlQQ/EYcMAH+4glVltdGB+YZHJ5tZr
+8BEchmxuQtYJtjahO1qkY3buHjiuQJsQCIZaQaXbKc5OoZBNHshttImtIwWYtR5c
+Df4KP97OA6TzPpx9ekL1rn9wJqogQmsMy6dqVvMkDfucheJwcIp2n2oaOdzVO3SO
+0+d6BmPLp5hnS9pBKFSxjXx/SsmHUJq4lcZmUOWHhRyn+lLfuJdY5ta93/axDlhA
+SMEhzpMwH65FLk0hpsk73bVR24uXP1TLqqaNRxUXnvB+KkawWkrX/PN81D8j+hON
+FsOmFp/xMukdkcBgoe85PkMDyBSa0FNMWY/AGW0bQ0YAbb1y4IVlzo1tLxxyYxZi
+pZvrPNOF1VRz3QVUvhs4
+-----END CERTIFICATE-----
