+++ driver/src/test/resources/multiple-sans.crt
@@ -1,0 +1,32 @@
+-----BEGIN CERTIFICATE-----
+MIIFkTCCA3mgAwIBAgIJAKBYC/K/0KcLMA0GCSqGSIb3DQEBCwUAMGIxCzAJBgNV
+BAYTAlVTMQswCQYDVQQIDAJDQTEQMA4GA1UEBwwHUGhvZW5peDEVMBMGA1UECgwM
+V2lkZ2V0IFdvcmtzMR0wGwYDVQQDDBR3d3cud2lkZ2V0LXdvcmtzLmNvbTAeFw0y
+MDEwMjYxNTU0MzZaFw0zMDEwMjQxNTU0MzZaMGIxCzAJBgNVBAYTAlVTMQswCQYD
+VQQIDAJDQTEQMA4GA1UEBwwHUGhvZW5peDEVMBMGA1UECgwMV2lkZ2V0IFdvcmtz
+MR0wGwYDVQQDDBR3d3cud2lkZ2V0LXdvcmtzLmNvbTCCAiIwDQYJKoZIhvcNAQEB
+BQADggIPADCCAgoCggIBAMopxJkgDSF43wPjI15HjY55DhRJlKte7ZXCyAf94uMj
+3xJyJBMxqoszDus/TG5HUnffyEWr4RQH1ZMSMmKZPWfZK664X1D3jnLm6/y9JVk6
+Tqsy/eqxTou4r1F5mjWj/X43vnON5bc6JAG3+QSp+iCAzd2Q6tMPScrXivuifjfx
+AkdMHxMYHTAouDfcFpHez/xDN9Z7LCmHgnnMKt6OjYC1nRCWIBwHjDzt1miq0xqY
+7F2pTzWKHksSBleR6R0Q286LQYzVYHdU4tVlTQpV+bHhcwprh/DamovXZvKnVEYy
+MH9IJV0uQm+J8zvanG7xUEnpoLikELDVsakgNgIIpHQAMvaAFvhz3C6CJ6zhyQvc
+lSusE2MqbEvv1pbOdcxENASeDx5LR+r+BoJPOwuyT9j5FV+UWnJZ2+pO1ObV+YIq
+QEcOr8aNTOGCNoPAFc45nWvUV2Epmjj97xHYdNqgPZdYBP49JZitClu9yigpO/bD
+wZB18zi296o9ajvTbfK5kewICPT6DaxOHZ8xYAZD6zr4WXZrJVOYVU5OooAHlhf5
+HMBn/2AdDxt2yr2hnbgFhgbqbTKYajF62T2CkvsvUkpxejx298Mu3rEnnTw+Iuky
+WBVHZ4bnGJLVT3jEc01K57SZsVwepu6X1X0GoeGKqi8DFGcW/X7kCwkgiY7DjiAd
+AgMBAAGjSjBIMEYGA1UdEQQ/MD2CEioud2lkZ2V0LXdvcmtzLm5ldIIVKi5lbnYu
+d2lkZ2V0LXdvcmtzLmlvghB3aWRnZXQtd29ya3MuY29tMA0GCSqGSIb3DQEBCwUA
+A4ICAQCIJSn04SfVpEq0yNpfqvzJm/Ry8Luz0XVbPVQfN16E/X/QMRck4kUb6jAy
+1tFKhmeO3hDpikvalDb4ieiqNurmQHGslVDL+8G6GpnRBVRVOaujmo0lloAaprVT
+3eI2qUgLfelACyDOPv7o2lrqWCL5UlExXOmL1hLpKUsCzxVZWfBy9cF3xGbd4Rob
+HnW36bmmyGe1+3HkAb+pXmGCUPi6AMdwQfsxoxeVr3WO5Yc/j5EC09tf5s3f3iB9
+U2b6WecqnpBi/kCwjPsYoe0+ZAgRIYMGmvkKbGKtB9fzpaZ3yilOPtbrCRRQifPf
+PlbL5buHlamI/rEW6pY9oYjwYV9Wll/JTcdv5uKAyk1UEyyV4xDJ1XbZqX20mLb2
+nhyw0UssxenjkUbysdiYUH5dBJFY9GEPcpDRuZ1J65Y50yGbea6SwN+JyHdHF6Ck
+jWhiXqW//k4ohlVHiJ+h7nQBgD4BS7klvweAivK2bdsJHY/qhyCUZ5RUvm82u5uF
+1WVCMTdgkdYBjJv7XOtu0vyD4bfdTH3ZYP/xmnbkj4EFoT6wywVUOr2zVDv7me1K
+pGMPdygWecFkPphXQAf+4XqNgqGtRwP5Fsm7Nos72oQQ8XcFTYQCn1mkqUYcoyLM
+vjkAZVnpp4riFp9oMUVcTss49Jxi4pJKPBLT2ODUHpp9CNOZqA==
+-----END CERTIFICATE-----
